Fixed a bug causing the high order double word of values returned by GetPerformanceCounter() became -1.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@1322 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdePkg/Library/BaseTimerLibLocalApic/x86TimerLib.c
@@ -210,7 +210,7 @@ GetPerformanceCounter (
   VOID
   )
 {
-  return InternalX86GetTimerTick (InternalX86GetApicBase ());
+  return (UINT32)InternalX86GetTimerTick (InternalX86GetApicBase ());
 }
 
 /**
